requirements.txt
console-menu
demoparser2
numpy
pandas
prettytable
python-dotenv
requests